chore(testing): harden coverage strategy and refactor integration suites
Raise test confidence by adding coverage thresholds/scripts, expanding unit and property coverage, splitting DynamoDB Local integration tests into focused suites with shared fixtures, and documenting the updated testing strategy.

Made-with: Cursor
