+++ .github/workflows/ci.yml
@@ -16,76 +16,11 @@ jobs:
       fail-fast: false
       matrix:
         node: [18, 20, 22]
-    runs-on: ubuntu-latest
-    timeout-minutes: 30
-    steps:
-      - uses: actions/checkout@v6
-
-      - uses: ./.github/actions/setup
-        with:
-          node-version: ${{ matrix.node }}
-
-      - name: Cache Turbo
-        uses: actions/cache@v4
-        with:
-          path: .turbo
-          key: turbo-${{ runner.os }}-${{ matrix.node }}-${{ github.sha }}
-          restore-keys: |
-            turbo-${{ runner.os }}-${{ matrix.node }}-
-            turbo-${{ runner.os }}-
-
-      - name: Start DynamoDB Local
-        run: docker compose up -d dynamodb-local
-
-      - name: Export DynamoDB endpoint
-        run: echo "DYNAMODB_ENDPOINT=http://localhost:8000" >> "$GITHUB_ENV"
-
-      - name: Wait for DynamoDB Local
-        run: |
-          for i in {1..30}; do
-            if (echo > /dev/tcp/localhost/8000) >/dev/null 2>&1; then
-              if curl -s -o /dev/null -w "%{http_code}" --max-time 2 http://localhost:8000 | grep -qE '^[1-5][0-9]{2}$'; then
-                echo "DynamoDB Local is accepting HTTP after ${i} attempts"
-                exit 0
-              fi
-            fi
-            sleep 2
-          done
-          echo "DynamoDB Local did not become ready in time"
-          docker compose logs dynamodb-local || true
-          exit 1
-
-      - name: Build
-        timeout-minutes: 10
-        run: pnpm build
-
-      - name: Generate API docs
-        run: pnpm docs:api
-
-      - name: Test
-        timeout-minutes: 15
-        run: pnpm test
-
-      - name: Format check
-        run: pnpm format:check
-
-      - name: Lint
-        run: pnpm lint
-
-      - name: Syncpack lint
-        run: pnpm syncpack lint
-
-      - name: Typecheck
-        run: pnpm typecheck
-
-      - name: publint
-        run: pnpm publint
-
-      - name: Are the Types Wrong
-        run: pnpm attw
-
-      - name: Smoke test tarballs
-        run: pnpm smoke:pack
+    uses: ./.github/workflows/validate.yml
+    with:
+      node-version: "${{ matrix.node }}"
+      turbo-cache-key-suffix: "${{ matrix.node }}"
+      coverage-artifact-name: "coverage-lcov-node-${{ matrix.node }}"
 
   dependency-review:
     if: github.event_name == 'pull_request'
