+++ package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "build": "turbo run build",
     "test": "turbo run test",
+    "test:coverage": "turbo run test:coverage",
     "lint": "eslint .",
     "typecheck": "turbo run typecheck",
     "format": "prettier --write .",
@@ -35,8 +36,10 @@
     "@commitlint/cli": "^20.5.0",
     "@commitlint/config-conventional": "^20.5.0",
     "@eslint/js": "^10.0.1",
+    "@vitest/coverage-v8": "3.2.4",
     "eslint": "^10.2.1",
     "eslint-config-prettier": "^10.1.8",
+    "fast-check": "^4.7.0",
     "husky": "^9.1.7",
     "lint-staged": "^16.4.0",
     "marked": "^18.0.2",
