+++ turbo.json
@@ -13,6 +13,12 @@
       "outputs": [],
       "env": ["DYNAMODB_ENDPOINT"]
     },
+    "test:coverage": {
+      "dependsOn": ["^build"],
+      "inputs": ["src/**", "test/**", "vitest.config.ts", "package.json"],
+      "outputs": ["coverage/**"],
+      "env": ["DYNAMODB_ENDPOINT"]
+    },
     "typecheck": {
       "dependsOn": ["^build"],
       "inputs": ["src/**", "test/**", "tsconfig.json", "package.json"],
